agent: handle 'zetcd', 'cetcd' database request

+++ agent/agent.go
@@ -68,7 +68,10 @@ var (
 
 	etcdBinaryPath   = filepath.Join(os.Getenv("GOPATH"), "bin/etcd")
 	consulBinaryPath = filepath.Join(os.Getenv("GOPATH"), "bin/consul")
-	javaBinaryPath   = "/usr/bin/java"
+	zetcdBinaryPath  = filepath.Join(os.Getenv("GOPATH"), "bin/zetcd")
+	cetcdBinaryPath  = filepath.Join(os.Getenv("GOPATH"), "bin/cetcd")
+
+	javaBinaryPath = "/usr/bin/java"
 
 	etcdToken     = "etcd_token"
 	etcdDataDir   = "data.etcd"
@@ -152,10 +155,15 @@ func CommandFunc(cmd *cobra.Command, args []string) error {
 }
 
 type transporterServer struct { // satisfy TransporterServer
-	req     Request
+	req Request
+
 	cmd     *exec.Cmd
 	logfile *os.File
 	pid     int
+
+	proxyCmd     *exec.Cmd
+	proxyLogfile *os.File
+	proxyPid     int
 }
 
 var uploadSig = make(chan Request_Operation, 1)
@@ -211,9 +219,15 @@ func (t *transporterServer) Transfer(ctx context.Context, r *Request) (*Response
 	switch r.Operation {
 	case Request_Start:
 		switch t.req.Database {
-		case Request_etcdv2, Request_etcdv3:
+		case Request_etcdv2, Request_etcdv3, Request_zetcd, Request_cetcd:
 			if !exist(etcdBinaryPath) {
-				return nil, fmt.Errorf("%q does not exist", etcdBinaryPath)
+				return nil, fmt.Errorf("etcd binary %q does not exist", etcdBinaryPath)
+			}
+			if t.req.Database == Request_zetcd && !exist(zetcdBinaryPath) {
+				return nil, fmt.Errorf("zetcd binary %q does not exist", zetcdBinaryPath)
+			}
+			if t.req.Database == Request_cetcd && !exist(cetcdBinaryPath) {
+				return nil, fmt.Errorf("cetcd binary %q does not exist", cetcdBinaryPath)
 			}
 			if err := os.RemoveAll(etcdDataDir); err != nil {
 				return nil, err
@@ -272,6 +286,47 @@ func (t *transporterServer) Transfer(ctx context.Context, r *Request) (*Response
 				}
 				plog.Infof("exiting %q", cmdString)
 			}()
+
+			if t.req.Database == Request_zetcd || t.req.Database == Request_cetcd {
+				f2, err := openToAppend(databaseLogPath + t.req.Database.String())
+				if err != nil {
+					return nil, err
+				}
+				t.proxyLogfile = f2
+				flags2 := []string{}
+				if t.req.Database == Request_zetcd {
+					flags2 = []string{
+						"-zkaddr", "0.0.0.0:2181",
+						"-endpoint", clientURLs[t.req.ServerIndex], // etcd endpoint
+					}
+				} else {
+					flags2 = []string{
+						"-consuladdr", "0.0.0.0:8500",
+						"-endpoint", clientURLs[t.req.ServerIndex], // etcd endpoint
+					}
+				}
+				flagString2 := strings.Join(flags2, " ")
+
+				cmd2 := exec.Command(zetcdBinaryPath, flags2...)
+				cmd2.Stdout = f2
+				cmd2.Stderr = f2
+
+				cmdString2 := fmt.Sprintf("%s %s", cmd2.Path, flagString2)
+				plog.Infof("starting binary %q", cmdString2)
+				if err := cmd2.Start(); err != nil {
+					return nil, err
+				}
+				t.proxyCmd = cmd2
+				t.proxyPid = cmd2.Process.Pid
+				plog.Infof("started binary %q [PID: %d]", cmdString2, t.proxyPid)
+				go func() {
+					if err := cmd2.Wait(); err != nil {
+						plog.Errorf("cmd.Wait %q returned error %v", cmdString2, err)
+						return
+					}
+					plog.Infof("exiting %q", cmdString2)
+				}()
+			}
 
 		case Request_ZooKeeper:
 			if !exist(javaBinaryPath) {
@@ -404,14 +459,14 @@ func (t *transporterServer) Transfer(ctx context.Context, r *Request) (*Response
 					plog.Error("cmd.Wait returned error", cmdString, err)
 					return
 				}
-				plog.Infof("exiting", cmdString, err)
+				plog.Infof("exiting %q (%v)", cmdString, err)
 			}()
 
 		default:
 			return nil, fmt.Errorf("unknown database %q", r.Database)
 		}
 
-	case Request_Restart:
+	case Request_Restart: // TODO: proxy is not supported!
 		if t.cmd == nil {
 			return nil, fmt.Errorf("nil command")
 		}
@@ -461,6 +516,15 @@ func (t *transporterServer) Transfer(ctx context.Context, r *Request) (*Response
 		}
 		if t.logfile != nil {
 			t.logfile.Close()
+		}
+		if t.proxyCmd != nil {
+			plog.Infof("stopping proxy binary %q [PID: %d]", t.req.Database.String(), t.proxyPid)
+			if err := syscall.Kill(t.proxyPid, syscall.SIGTERM); err != nil {
+				return nil, err
+			}
+		}
+		if t.proxyLogfile != nil {
+			t.proxyLogfile.Close()
 		}
 		plog.Infof("stopped binary %q [PID: %d]", t.req.Database.String(), t.pid)
 		processPID = t.pid
